Update mData_2 variable location from GraphAttributeWidget.cpp file to BrainAliveBluetooth.cpp file and fix few bugs
